Added exception handling and proper messages returned from dao layer.

+++ database/src/main/java/com/iquestgroup/database/impl/DefaultAuthorDAO.java
@@ -1,11 +1,10 @@
 package com.iquestgroup.database.impl;
 
 import com.iquestgroup.database.AuthorDAO;
+import com.iquestgroup.database.exceptionHandling.DAOException;
 import com.iquestgroup.model.Author;
 import org.hibernate.*;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -15,7 +14,7 @@ public class DefaultAuthorDAO implements AuthorDAO {
     private SessionFactory sessionFactory;
 
     @Override
-    public List<Author> getAllAuthors() {
+    public List<Author> getAllAuthors() throws DAOException {
         List<Author> resultList = new ArrayList<>();
         Transaction transaction = null;
 
@@ -25,18 +24,18 @@ public class DefaultAuthorDAO implements AuthorDAO {
             for (Object author : authorList) {
                 resultList.add((Author) author);
             }
-        } catch (HibernateError e) {
+        } catch (HibernateException e) {
             if (transaction != null) {
                 transaction.rollback();
             }
-            e.printStackTrace();
+            throw new DAOException("An error occurred while trying to retrieve all authors from the database", e);
         }
 
         return resultList;
     }
 
     @Override
-    public Author getAuthorByID(Integer authorID) {
+    public Author getAuthorByID(Integer authorID) throws DAOException {
         Author author = null;
         Transaction transaction = null;
 
@@ -48,14 +47,16 @@ public class DefaultAuthorDAO implements AuthorDAO {
             if (transaction != null) {
                 transaction.rollback();
             }
-            e.printStackTrace();
+            throw new DAOException(
+                "An error occurred while trying to retrieve the author with the id " + authorID + "from the database",
+                e);
         }
 
         return author;
     }
 
     @Override
-    public void insertAuthor(Author author) {
+    public String insertAuthor(Author author) throws DAOException {
         Transaction transaction = null;
 
         try (Session session = sessionFactory.openSession()) {
@@ -66,24 +67,34 @@ public class DefaultAuthorDAO implements AuthorDAO {
             if (transaction != null) {
                 transaction.rollback();
             }
-            e.printStackTrace();
+            throw new DAOException("An error occured while trying to insert " + author, e);
         }
+
+        return "Author with name " + author.getName() + " successfully inserted!";
     }
 
     @Override
-    public void deleteAuthor(Integer authorID) {
+    public String deleteAuthor(Integer authorID) throws DAOException {
         Transaction transaction = null;
 
         try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
             Author author = session.get(Author.class, authorID);
-            session.delete(author);
-            transaction.commit();
+
+            if (author == null) {
+                return "Author with id: " + authorID + " does not exist in the database!";
+            } else {
+                session.delete(author);
+                transaction.commit();
+            }
         } catch (HibernateException e) {
             if (transaction != null) {
                 transaction.rollback();
             }
-            e.printStackTrace();
+            throw new DAOException(
+                "An error occured while trying to delete author with id: " + authorID + " from the database", e);
         }
+
+        return "Author with id: " + authorID + " successfully deleted from the database!";
     }
 }
